+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Further compatibility fixes by using `TypeMirror.getKind()` rather than `instanceof SomeType`.
+* Change the `SuppressWarningsUtil` code so that id a class is a nested class and one of the outer classes are deprecated then the class is considered deprecated.
 
 ### [v0.27](https://github.com/realityforge/proton/tree/v0.27) (2020-01-29) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.26...v0.27)